+++ unit-tests/json-to-iac/power-vs.test.js
@@ -730,5 +730,238 @@ resource "ibm_pi_cloud_connection_network_attach" "power_example_dev_connection_
         "it should return correctly formatted data"
       );
     });
+    it("should return a power vs network configuration with with multiple images and networks", () => {
+      let config = {
+        _options: {
+          tags: ["hello", "world"],
+        },
+        resource_groups: [
+          {
+            use_data: false,
+            name: "example",
+          },
+        ],
+        power: [
+          {
+            name: "example",
+            resource_group: "example",
+            zone: "dal10",
+            ssh_keys: [
+              {
+                workspace: "example",
+                name: "keyname",
+                zone: "dal10",
+              },
+            ],
+            network: [
+              {
+                workspace: "example",
+                name: "dev-nw",
+                pi_cidr: "1.2.3.4/5",
+                pi_dns: ["127.0.0.1"],
+                pi_network_type: "vlan",
+                pi_network_jumbo: true,
+                zone: "dal10",
+              },
+              {
+                workspace: "example",
+                name: "dev-nw2",
+                pi_cidr: "1.2.3.4/5",
+                pi_dns: ["127.0.0.1"],
+                pi_network_type: "vlan",
+                pi_network_jumbo: true,
+                zone: "dal10",
+              },
+            ],
+            cloud_connections: [
+              {
+                name: "dev-connection",
+                workspace: "example",
+                pi_cloud_connection_speed: 50,
+                pi_cloud_connection_global_routing: false,
+                pi_cloud_connection_metered: false,
+                pi_cloud_connection_transit_enabled: true,
+                transit_gateways: [],
+                zone: "dal10",
+              },
+            ],
+            images: [
+              {
+                workspace: "example",
+                pi_image_id: "e4de6683-2a42-4993-b702-c8613f132d39",
+                name: "SLES15-SP3-SAP",
+                zone: "dal10",
+              },
+              {
+                workspace: "example",
+                pi_image_id: "e4de6683-2a42-4993-b702-c8613f132d39",
+                name: "SLES15-SP3-SAP2",
+                zone: "dal10",
+              },
+            ],
+            attachments: [
+              {
+                connections: ["dev-connection"],
+                workspace: "example",
+                network: "dev-nw",
+                zone: "dal10",
+              },
+            ],
+          },
+        ],
+      };
+      let actualData = powerVsTf(config);
+      let expectedData = `##############################################################################
+# Power VS Workspace Example
+##############################################################################
+
+resource "ibm_resource_instance" "power_vs_workspace_example" {
+  provider          = ibm.power_vs_dal10
+  name              = "\${var.prefix}-power-workspace-example"
+  service           = "power-iaas"
+  plan              = "power-virtual-server-group"
+  location          = "dal10"
+  resource_group_id = ibm_resource_group.example.id
+  tags = [
+    "hello",
+    "world"
+  ]
+  timeouts {
+    create = "6m"
+    update = "5m"
+    delete = "10m"
+  }
+}
+
+##############################################################################
+
+##############################################################################
+# Example Workspace SSH Keys
+##############################################################################
+
+resource "ibm_pi_key" "power_vs_ssh_key_keyname" {
+  provider             = ibm.power_vs_dal10
+  pi_cloud_instance_id = ibm_resource_instance.power_vs_workspace_example.guid
+  pi_key_name          = "\${var.prefix}-power-example-keyname-key"
+  pi_ssh_key           = var.power_example_keyname_key
+}
+
+##############################################################################
+
+##############################################################################
+# Example Workspace Network
+##############################################################################
+
+resource "ibm_pi_network" "power_network_example_dev_nw" {
+  provider             = ibm.power_vs_dal10
+  pi_cloud_instance_id = ibm_resource_instance.power_vs_workspace_example.guid
+  pi_network_name      = "\${var.prefix}-power-network-dev-nw"
+  pi_cidr              = "1.2.3.4/5"
+  pi_network_type      = "vlan"
+  pi_network_jumbo     = true
+  pi_dns = [
+    "127.0.0.1"
+  ]
+}
+
+resource "ibm_pi_network" "power_network_example_dev_nw2" {
+  provider             = ibm.power_vs_dal10
+  pi_cloud_instance_id = ibm_resource_instance.power_vs_workspace_example.guid
+  pi_network_name      = "\${var.prefix}-power-network-dev-nw2"
+  pi_cidr              = "1.2.3.4/5"
+  pi_network_type      = "vlan"
+  pi_network_jumbo     = true
+  pi_dns = [
+    "127.0.0.1"
+  ]
+  depends_on = [
+    ibm_pi_network.power_network_example_dev_nw
+  ]
+}
+
+##############################################################################
+
+##############################################################################
+# Example Workspace Images
+##############################################################################
+
+resource "ibm_pi_image" "power_image_example_sles15_sp3_sap" {
+  provider             = ibm.power_vs_dal10
+  pi_cloud_instance_id = ibm_resource_instance.power_vs_workspace_example.guid
+  pi_image_id          = "e4de6683-2a42-4993-b702-c8613f132d39"
+  pi_image_name        = "SLES15-SP3-SAP"
+  timeouts {
+    create = "9m"
+  }
+}
+
+resource "ibm_pi_image" "power_image_example_sles15_sp3_sap2" {
+  provider             = ibm.power_vs_dal10
+  pi_cloud_instance_id = ibm_resource_instance.power_vs_workspace_example.guid
+  pi_image_id          = "e4de6683-2a42-4993-b702-c8613f132d39"
+  pi_image_name        = "SLES15-SP3-SAP2"
+  timeouts {
+    create = "9m"
+  }
+  depends_on = [
+    ibm_pi_image.power_image_example_sles15_sp3_sap
+  ]
+}
+
+##############################################################################
+
+##############################################################################
+# Example Workspace Dev Connection
+##############################################################################
+
+resource "ibm_pi_cloud_connection" "power_network_example_connection_dev_connection" {
+  provider                            = ibm.power_vs_dal10
+  pi_cloud_instance_id                = ibm_resource_instance.power_vs_workspace_example.guid
+  pi_cloud_connection_name            = "\${var.prefix}-power-network-dev-connection-connection"
+  pi_cloud_connection_speed           = 50
+  pi_cloud_connection_global_routing  = false
+  pi_cloud_connection_metered         = false
+  pi_cloud_connection_transit_enabled = true
+}
+
+data "ibm_dl_gateway" "power_network_example_connection_dev_connection" {
+  provider = ibm.power_vs_dal10
+  name     = "\${var.prefix}-power-network-dev-connection-connection"
+  depends_on = [
+    ibm_pi_cloud_connection.power_network_example_connection_dev_connection
+  ]
+}
+
+resource "time_sleep" "power_network_example_connection_dev_connection_sleep" {
+  create_duration = "120s"
+  triggers = {
+    crn = data.ibm_dl_gateway.power_network_example_connection_dev_connection.crn
+  }
+  depends_on = [
+    data.ibm_dl_gateway.power_network_example_connection_dev_connection
+  ]
+}
+
+##############################################################################
+
+##############################################################################
+# Example Workspace Network Attachments
+##############################################################################
+
+resource "ibm_pi_cloud_connection_network_attach" "power_example_dev_connection_connection_dev_nw_connection" {
+  provider               = ibm.power_vs_dal10
+  pi_cloud_instance_id   = ibm_resource_instance.power_vs_workspace_example.guid
+  pi_cloud_connection_id = ibm_pi_cloud_connection.power_network_example_connection_dev_connection.cloud_connection_id
+  pi_network_id          = ibm_pi_network.power_network_example_dev_nw.network_id
+}
+
+##############################################################################
+`;
+      assert.deepEqual(
+        actualData,
+        expectedData,
+        "it should return correctly formatted data"
+      );
+    });
   });
 });